Add PUT/:id api endpoint
In this commit, I added a new api endpoint for updating an account record.
This endpoint expects a name and budget fields within the req.body of the
request. It then uses the payload to replace the account record which matches
the specified id (req.params.id) within the database